util/util-jvm: Add gc pause stats for all collectors
Problem

Currently we only have gc pause stats for the ParNew collector, but
we'd like to be able to see this information for the G1 pools.

Solution

Add collection duration stats for all collector pools. Keep the jvm/gc/eden/pause_msec
stat for compatibility, but this will be the same as jvm/gc/ParNew/pause_msec.

Differential Revision: https://phabricator.twitter.biz/D1176049