finagle/finagle-netty4: Add CPU tracking and active socket count for worker threads in EventLoopGroupTracker
Problem

We'd like to be able to see how much work different netty threads are doing,
in order to size the pool correctly and potentially improve work distribution.

Solution

As an initial step, add stats for the cpu time for each thread and the number of
active sockets it's assigned to EventLoopGroupTracker. These stats will only be
collected if the `TrackWorkerPool.enableTracking` param is set to `true` on the
server (false by default).

Differential Revision: https://phabricator.twitter.biz/D1176906